requirements.txt
asgiref==3.8.1
Django==5.0.4
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.4
django-js-asset==2.2.0
django-livereload-server==0.5.1
pillow==10.3.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
tornado==6.4
